display: print cluster info from metadata in profile dir
