+++ pkg/utils/tui.go
@@ -14,6 +14,9 @@
 package utils
 
 import (
+	"strconv"
+	"strings"
+
 	"github.com/cheynewallace/tabby"
 )
 
@@ -44,4 +47,14 @@ func addRow(t *tabby.Tabby, rawLine []string, header bool) {
 	} else {
 		t.AddLine(row...)
 	}
+}
+
+// JoinInt joins a slice of int to string
+func JoinInt(nums []int, delim string) string {
+	result := ""
+	for _, i := range nums {
+		result += strconv.Itoa(i)
+		result += delim
+	}
+	return strings.TrimSuffix(result, delim)
 }
